Started working on the filter.
There's still some work to do, is not giving the desired results. We need to get the right side of the disparity for it to work better. As of now I'm just passing the same disparity as left and right. Because for some reason even tho the documentation says that the right disparity is optional it will yield an Assertion fault error if there's none present. :(